// src/types/assignment.ts
export interface Assignment {
  id: number;
  customer_id: number;
  customer_first_name: string;
  customer_last_name: string;
  coordinator_id: number;
  coordinator_name: string;
  coordinator_email: string;
  sales_id: number;
  sales_name: string;
  sales_email: string;
  assignment_date: string;
  status: number;
}